+++ kernel/task.h
@@ -31,7 +31,7 @@ err_t task_fork(u64 *pid);
 void task_legacy_switch(void);
 void task_exec(struct sv file);
 err_t task_fd_open(u64 *fd, struct sv path, int flags);
-err_t task_fd_write(int fd, const void *buffer, u64 count, u64 *out);
+err_t task_fd_write(u64 fd, const void *buffer, u64 count, u64 *out);
 void task_fd_close(u64 fd);
 void *task_sbrk(uintptr_t increment);
 #endif // TASK_H
